crud de dividas pronto

+++ financeiro/models.py
@@ -59,7 +59,7 @@ class Caixinhas(models.Model):
 class Dividas(models.Model):
     nome = models.CharField(max_length=100)
     valor_parcela = models.DecimalField(max_digits=10, decimal_places=2)
-    total_parcela = models.DecimalField(max_digits=10, decimal_places=2)
+    total_parcela = models.IntegerField()
     data = models.DateField()
 
     def __str__(self):
